requirements.txt
pymysql==0.9.3
mysqlclient>=1.4.4
Scrapy>=1.7.3
shub>=2.9.0
SQLAlchemy>=1.3.6
scrapy-useragents==0.0.1
scrapy-rotating-proxies==0.6.2
selenium==3.141.0
scrapy-selenium==0.0.7
psycopg2-binary==2.8.5